requirements.txt
numpy
Cython
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==0.22.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
